+++ scenario-compiler/src/main/scala/pl/touk/nussknacker/engine/definition/action/ModelDataActionInfoProvider.scala
@@ -4,7 +4,7 @@ import cats.data.ValidatedNel
 import com.typesafe.scalalogging.LazyLogging
 import pl.touk.nussknacker.engine.ModelData
 import pl.touk.nussknacker.engine.api.{JobData, ProcessVersion}
-import pl.touk.nussknacker.engine.api.component.{NodeComponentInfo, ParameterConfig}
+import pl.touk.nussknacker.engine.api.component.{NodeComponentInfo, StaticParameterConfig}
 import pl.touk.nussknacker.engine.api.context.ProcessCompilationError
 import pl.touk.nussknacker.engine.api.deployment.{ScenarioActionName, WithActionParametersSupport}
 import pl.touk.nussknacker.engine.api.parameter.ParameterName
@@ -19,7 +19,7 @@ class ModelDataActionInfoProvider(modelData: ModelData) extends ActionInfoProvid
       scenario: CanonicalProcess
   ): ValidatedNel[
     ProcessCompilationError,
-    Map[ScenarioActionName, Map[NodeComponentInfo, Map[ParameterName, ParameterConfig]]]
+    Map[ScenarioActionName, Map[NodeComponentInfo, Map[ParameterName, StaticParameterConfig]]]
   ] = {
     val jobData = JobData(scenario.metaData, processVersion)
     modelData.withThisAsContextClassLoader {
@@ -31,7 +31,7 @@ class ModelDataActionInfoProvider(modelData: ModelData) extends ActionInfoProvid
 
   private def extractParametersFromCustomNodes(
       compiledCustomNodes: Map[NodeComponentInfo, Any]
-  ): Map[ScenarioActionName, Map[NodeComponentInfo, Map[ParameterName, ParameterConfig]]] = {
+  ): Map[ScenarioActionName, Map[NodeComponentInfo, Map[ParameterName, StaticParameterConfig]]] = {
     val nodeToActionToParameters = compiledCustomNodes
       .mapValuesNow {
         case s: WithActionParametersSupport => Some(s.actionParametersDefinition)
@@ -44,10 +44,10 @@ class ModelDataActionInfoProvider(modelData: ModelData) extends ActionInfoProvid
   }
 
   private def groupByAction(
-      nodeToActionToParameters: Map[NodeComponentInfo, Map[ScenarioActionName, Map[ParameterName, ParameterConfig]]]
-  ): Map[ScenarioActionName, Map[NodeComponentInfo, Map[ParameterName, ParameterConfig]]] = {
+      nodeToActionToParams: Map[NodeComponentInfo, Map[ScenarioActionName, Map[ParameterName, StaticParameterConfig]]]
+  ): Map[ScenarioActionName, Map[NodeComponentInfo, Map[ParameterName, StaticParameterConfig]]] = {
     val actionToNodeToParameters = for {
-      (node, actionToParams) <- nodeToActionToParameters.toList
+      (node, actionToParams) <- nodeToActionToParams.toList
       (actionName, params)   <- actionToParams.toList
     } yield (actionName, node -> params)
     actionToNodeToParameters